pyproject.toml
[project]
name = "fastapi-production-template"
version = "1.0.0"
description = "Production-ready async FastAPI template with PostgreSQL, circuit breakers, RabbitMQ, Kafka, ELK, Prometheus, and LangChain/LangGraph support."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "tenacity>=9.0.0",
    "aiobreaker>=1.2.0",
    "prometheus-client>=0.21.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "aio-pika>=9.4.0",
    "aiokafka>=0.11.0",
]

[project.optional-dependencies]
ai = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
]
elk = [
    "elasticsearch>=8.13.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "coverage>=7.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
